requirements.txt
fastapi
uvicorn
Pillow
opencv-python-headless
torch
ultralytics